more test cleanup plus test api integration test
