+++ src/NetDaemon.HassClient.Tests/Integration/Testdata/result_get_entities.json
@@ -8,6 +8,7 @@
             "device_id": "42cdda32a2a3428e86c2e27699d79ead",
             "disabled_by": null,
             "entity_id": "media_player.tv_uppe2",
+            "area_id": "42cdda1212a3428e86c2e27699d79ead",
             "name": null,
             "icon": null,
             "platform": "cast"
