+++ src/NetDaemon.HassClient.Tests/Integration/WebsocketIntegrationTests.cs
@@ -1,31 +1,10 @@
 namespace NetDaemon.HassClient.Tests.Integration;
 
-public class HomeAssistantServiceFixture : IAsyncLifetime
+public class WebsocketIntegrationTests : IntegrationTestBase
 {
-    public HomeAssistantMock? HaMock { get; set; }
-    public async Task DisposeAsync()
+    public WebsocketIntegrationTests(HomeAssistantServiceFixture fixture) : base(fixture)
     {
-        if (HaMock is not null)
-            await HaMock.DisposeAsync().ConfigureAwait(false);
     }
-
-    public Task InitializeAsync()
-    {
-        HaMock = new HomeAssistantMock();
-        return Task.CompletedTask;
-    }
-}
-
-public class IntegrationTests : IClassFixture<HomeAssistantServiceFixture>
-{
-    public CancellationTokenSource _tokenSource = new(TestSettings.DefaultTimeout);
-    public IntegrationTests(HomeAssistantServiceFixture fixture)
-    {
-        HaFixture = fixture;
-    }
-
-    private HomeAssistantServiceFixture HaFixture { get; }
-
 
     [Fact]
     public async Task TestSuccessfulConnectShouldReturnConnection()
@@ -215,6 +194,10 @@ public class IntegrationTests : IClassFixture<HomeAssistantServiceFixture>
             .BatteryLevel
             .Should()
             .Be(100);
+
+        changedEvent?.NewState?.Attributes?.FirstOrDefault(n => n.Key == "battery_level")
+            .Should()
+            .NotBeNull();
     }
 
     [Fact]
@@ -253,68 +236,4 @@ public class IntegrationTests : IClassFixture<HomeAssistantServiceFixture>
             .BeEquivalentTo("light");
     }
 
-    private record TestContext : IAsyncDisposable
-    {
-        public Mock<ILogger<HomeAssistantClient>> HomeAssistantLogger { get; init; } = new();
-        public Mock<ILogger<IWebSocketClientTransportPipeline>> TransportPipelineLogger { get; init; } = new();
-        public Mock<ILogger<IHomeAssistantConnection>> HomeAssistantConnectionLogger { get; init; } = new();
-        public IHomeAssistantConnection HomeAssistantConnction { get; init; } = new Mock<IHomeAssistantConnection>().Object;
-        public async ValueTask DisposeAsync()
-        {
-            await HomeAssistantConnction.DisposeAsync().ConfigureAwait(false);
-        }
-    }
-
-    /// <summary>
-    ///     Returns a connection Home Assistant instance
-    /// </summary>
-    /// <param name="haSettings">Provide custom setting</param>
-    private async Task<TestContext> GetConnectedClientContext(HomeAssistantSettings? haSettings = null)
-    {
-        var mock = HaFixture?.HaMock ?? throw new ApplicationException("Unexpected for the mock server to be null");
-
-        var loggerClient = new Mock<ILogger<HomeAssistantClient>>();
-        var loggerTransport = new Mock<ILogger<IWebSocketClientTransportPipeline>>();
-        var loggerConnection = new Mock<ILogger<IHomeAssistantConnection>>();
-
-        var settings = haSettings ?? new HomeAssistantSettings()
-        {
-            Host = "127.0.0.1",
-            Port = mock?.ServerPort ?? 0,
-            Ssl = false,
-            Token = "ABCDEFGHIJKLMNOPQ"
-        };
-
-        IOptions<HomeAssistantSettings> appSettingsOptions = Options.Create(settings);
-
-        var client = new HomeAssistantClient(
-            loggerClient.Object,
-            new WebSocketClientFactory(),
-            new WebSocketClientTransportPipelineFactory(loggerTransport.Object),
-            new HomeAssistantConnectionFactory(
-                loggerConnection.Object,
-                new HomeAssistantApiManager(
-                    appSettingsOptions,
-                    (mock?.HomeAssistantHost.Services.GetRequiredService<IHttpClientFactory>() ?? throw new NullReferenceException())
-                    .CreateClient()
-                )
-            )
-
-        );
-        var connection = await client.ConnectAsync(
-            settings.Host,
-            settings.Port,
-            settings.Ssl,
-            settings.Token,
-            _tokenSource.Token
-        ).ConfigureAwait(false);
-
-        return new TestContext
-        {
-            HomeAssistantLogger = loggerClient,
-            TransportPipelineLogger = loggerTransport,
-            HomeAssistantConnectionLogger = loggerConnection,
-            HomeAssistantConnction = connection
-        };
-    }
 }
